Move message/tool types to a torch-free tinker_cookbook.tool_types module
`tinker_cookbook.renderers.base` imports `tinker`/`torch` for renderer logic,
so anything importing the message/tool types (`Message`, `ToolCall`, `ToolSpec`,
content parts) through `tinker_cookbook.renderers` transitively pulls in torch.
That breaks lightweight, torch-free consumers — e.g. an HTTP tool server that
only needs the type definitions and runs on a CPU-only node, where CUDA torch
fails to import (`libcudnn.so not found`).

Move the (torch-free) type definitions into a new top-level
`tinker_cookbook.tool_types` module. `renderers/base.py` imports and re-exports
them, so the existing `from tinker_cookbook.renderers import ToolCall` API is
unchanged. Consumers that want to avoid torch can now
`from tinker_cookbook.tool_types import ToolCall, Message, ToolSpec`, which
only triggers the torch-free package `__init__` (verified: importing it loads
neither `torch` nor `tinker`).
